Checkpoint: Paths are now properly configured
