+++ tests/cart_test.cpp
@@ -22,6 +22,7 @@
 
 #include "bus.h"
 #include "cartridge.h"
+#include "config.h"
 #include <fmt/base.h>
 #include <gtest/gtest.h>
 
@@ -36,7 +37,7 @@ class CartTest : public ::testing::Test
 
     CartTest()
     {
-        std::string romFile = "tests/roms/palette.nes";
+        std::string romFile = std::string( ROM_DIR ) + "/palette.nes";
         bus.cartridge.LoadRom( romFile );
         bus.cpu.Reset();
     }
