--- tools/cli/example.cpp
@@ -1,43 +1,0 @@
-#include <pybind11/pybind11.h>
-
-#define STRINGIFY( x ) #x
-#define MACRO_STRINGIFY( x ) STRINGIFY( x )
-
-int add( int i, int j )
-{
-    return i + j;
-}
-
-PYBIND11_MODULE( cli_module, m )
-{
-    m.doc() = R"pbdoc(
-        Pybind11 example plugin
-        -----------------------
-
-        .. currentmodule:: cli_module
-
-        .. autosummary::
-           :toctree: _generate
-
-           add
-           subtract
-    )pbdoc";
-
-    m.def( "add", &add, R"pbdoc(
-        Add two numbers
-
-        Some other explanation about the add function.
-    )pbdoc" );
-
-    m.def( "subtract", []( int i, int j ) { return i - j; }, R"pbdoc(
-        Subtract two numbers
-
-        Some other explanation about the subtract function.
-    )pbdoc" );
-
-#ifdef VERSION_INFO
-    m.attr( "__version__" ) = MACRO_STRINGIFY( VERSION_INFO );
-#else
-    m.attr( "__version__" ) = "dev";
-#endif
-}
