--- tools/cli/scripts/build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-cd "$(dirname "$BASH_SOURCE[0]")/../../.." || exit 1
-scripts/build.sh cli
