+++ tools/python/scripts/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd "$(dirname "$BASH_SOURCE[0]")/../../.." || exit 1
+scripts/build.sh python
